Update README.md
A Space Odyssey - "2001: A Space Odyssey" is a groundbreaking science fiction film directed by Stanley Kubrick. Released in 1968, it is known for its stunning visuals and philosophical themes.

Citizen Kane - Directed by Orson Welles and released in 1941, "Citizen Kane" is often regarded as one of the greatest films in the history of cinema. It revolutionized filmmaking with its innovative techniques.

La Dolce Vita - Directed by Federico Fellini and released in 1960, "La Dolce Vita" is an Italian film that explores the decadence and hedonism of Roman high society.

In the Mood for Love - A Hong Kong romantic drama film directed by Wong Kar-wai, released in 2000. It is praised for its lush cinematography and emotionally resonant storytelling.

There Will Be Blood - Directed by Paul Thomas Anderson and released in 2007, this film is a powerful drama starring Daniel Day-Lewis. It explores the ruthless world of early 20th-century oil prospecting.

Singin' in the Rain - A classic musical comedy released in 1952, directed by Stanley Donen and Gene Kelly. It celebrates the transition from silent films to "talkies."

GoodFellas - Directed by Martin Scorsese and released in 1990, this crime film is based on the true story of Henry Hill's life in the mob. It is widely regarded as one of the greatest films in the gangster genre.

North by Northwest - A classic Hitchcock thriller released in 1959, starring Cary Grant and Eva Marie Saint. It's known for its iconic scenes, including the famous crop duster sequence.

The Office - This refers to the American version of "The Office," a mockumentary sitcom that ran from 2005 to 2013. It was created by Greg Daniels and became widely popular for its unique format and humor.